requirements.txt
Django>=4.1
psycopg2-binary>=2.8
django-phonenumber-field[phonenumberslite]
django-countries==7.4.2
django-localflavor~=3.1
djangorestframework~=3.14.0
django-scheduler~=0.10.0
pandas